import wrapped mantid algorithms on launch
