+++ mslice/cli/_mslice_commmands.py
@@ -11,8 +11,10 @@ import os.path as ospath
 from mantid.api import IMDWorkspace as _IMDWorkspace
 
 from mslice.app import MAIN_WINDOW
+
 from mslice.workspace.base import WorkspaceBase as Workspace
-from mslice.util.mantid import run_algorithm, Scale
+from mslice.util.mantid.algorithm_wrapper import run_algorithm, Scale
+
 # Helper tools
 from mslice.models.workspacemanager.workspace_provider import get_workspace_handle, workspace_exists
 from mslice.presenters.slice_plotter_presenter import Axis as _Axis
